Bug #378054 - modify .ko to build-in kernel and revert armlog modify in engpc
[bug number  ] 378054
[root cause  ] modify .ko to build-in kernel and revert armlog modify in engpc
[changes     ] download sdiodevhal mdbg
[side effects] no
[self test   ] [board:sp9630ea3mn] [teset case:1.BT/WIFI function normal 2.armlog and slog normal ]
[whether AOB ]
[reviewers   ] paul.luo gaole.zhang

+++ drivers/misc/mdbg/mdbg_type.h
@@ -16,11 +16,9 @@
 #include <linux/miscdevice.h>
 #include <linux/delay.h>
 #include <linux/module.h>
-#include <linux/interrupt.h>
 #include <linux/slab.h>
 #include <linux/irq.h>
 #include <linux/input.h>
-#include <linux/gpio.h>
 #include <linux/fs.h>
 #include <asm/uaccess.h>
 #include <linux/miscdevice.h>
@@ -61,7 +59,7 @@
 #define MDBG_ERR_NO_FILE		(-6)
 
 
-typedef int MDBG_SIZE_T;
+typedef long int MDBG_SIZE_T;
 
 
 #endif
